Streamlined and fixed scripts, added customized events
Removed redundant variables in favor of events (in toolmanager and inputlistener), that pass the required information when needed. Improved inputlistener getting correct indexes AFTER the game has already started and added suitable events for it. Laserpointers and drawing now work with toolmanager correctly when networked, functionality not tested with multiple people yet